Module 2/FPTSupportSystem/backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# LangChain / LangGraph
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# RAG - Vector Store
chromadb>=0.4.22
pypdf>=4.0.0

# Web Search
tavily-python>=0.3.0

# HTTP Client
httpx>=0.26.0

# CORS
python-multipart>=0.0.6
